image-detector/dfine/requirements.txt
torch==2.6.0
transformers>=4.51.1
pillow==10.4.0
requests>=2.32.3
timm==1.0.12
opencv-python-headless==4.10.0.84
clarifai>=12.1.3
hf_transfer
